requirements.txt
asgiref==3.6.0
autopep8==2.0.2
bcrypt==4.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.1.7
django-configurations==2.4
djangorestframework==3.14.0
idna==3.4
psycopg2==2.9.5
pycodestyle==2.10.0
PyJWT==2.6.0
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.30.0
sqlparse==0.4.4
tomli==2.0.1
tzdata==2022.7
urllib3==2.0.2
